Scheduling (#6, PR #36)
Scheduling is implemented for both Kubernetes and Docker. When scheduling is not enabled (`max_proc` is not overridden), scrapyd-k8s has the old behaviour, without any scheduling. Only when it is enabled, the maximum number of processes is limited, and a new job is only started when there is an emtpy slot.

Note that for Kubernetes, the existing resource watcher is used. For Docker, a thread polls the status every 5 seconds.

The testing infrastructure was expanded, so that multiple configuration files can be tested, as well as test-specific Kubernetes setup (in this case, to add and remove API permissions for specific tests).

Note that this also introduces unit-tests for Kubernetes and Docker scheduling. This was good to test, but the main testing approach for scrapyd-k8s remains integration tests, where the tests are separate from the scrapyd-k8s instance under test.

---------

Co-authored-by: wvengen <willem@thequestionmark.org>